requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-google-genai
faiss-cpu  # Required for FAISS vector storage
google-generativeai
tiktoken  # Tokenizer required for LangChain
